Iterative Schema with TFX and ML Metadata
In this notebook, you will review how to update an inferred schema and save the result to the metadata store used by TFX. The TFX components get information from this database before executing executions. Thus, if you curate a schema, you will need to save this as an artifact in the metadata store. You will get to see how that is done in the following exercise.

Afterwards, you will also practice accessing the TFX metadata store and see how you can track the lineage of an artifact.