Distributed training using Mirrored strategy
A Keras model that was designed to run on a single worker can seamlessly work on multiple workers with the minimal code change. In particular, the notebook will:
Perform training with a single worker.
Understand the requirements for a multi-worker setup (tf_config variable) and use context managers for implementing distributed strategies.
Use magic commands to simulate different machines.
Perform a multi-worker training strategy.